requirements.txt
certifi==2020.6.20
chardet==3.0.4
httpbl==1.0.1
idna==2.10
pyaml==20.4.0
PyYAML==5.3.1
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
urllib3==1.25.10
